One small change to the Protocol and nice comments in the example imp…

+++ AiProtocol/AiProtocol/Commands.cs
@@ -40,11 +40,11 @@ namespace AiProtocol.Command
     public class Movement : BaseCommand
     {
         public MovementStyle Style;
-        public int xCoord;
-        public int yCoord;
+        public float xCoord;
+        public float yCoord;
 
         public Movement() { }
-        public Movement(MovementStyle s, int x, int y)
+        public Movement(MovementStyle s, float x, float y)
         {
             Style = s;
             xCoord = x;
